# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards API
  description: |-
    REST API системы управления банковскими картами.

        Функциональность:
        - регистрация и аутентификация пользователей
        - управление банковскими картами
        - переводы между картами
        - заявки на активацию и блокировку карт
        - разграничение доступа по ролям (ADMIN / USER)

        Аутентификация осуществляется с помощью JWT (Bearer Token).
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/card-requests/{cardRequestId}/reject:
    post:
      tags:
      - Card Requests
      summary: Отклонить заявку (ADMIN)
      operationId: reject
      parameters:
      - name: cardRequestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: No Content
      security:
      - bearerAuth: []
  /api/card-requests/{cardRequestId}/approve:
    post:
      tags:
      - Card Requests
      summary: Одобрить заявку (ADMIN)
      operationId: approve
      parameters:
      - name: cardRequestId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: No Content
      security:
      - bearerAuth: []
  /api/card-requests/block:
    post:
      tags:
      - Card Requests
      summary: Создать заявку на блокировку карты
      description: Пользователь отправляет запрос на блокировку своей карты
      operationId: createBlock
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRequestBlockDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав USER
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Заявка создана
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardRequestDto"
      security:
      - bearerAuth: []
  /api/card-requests/activate:
    post:
      tags:
      - Card Requests
      summary: Создать заявку на активацию карты
      description: Пользователь отправляет запрос на активацию своей карты
      operationId: createActivate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardRequestActivateDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав USER
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Заявка создана
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardRequestDto"
      security:
      - bearerAuth: []
  /api/auth/refresh:
    post:
      tags:
      - Auth
      summary: Обновление access token
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtAuthenticationDto"
  /api/auth/login:
    post:
      tags:
      - Auth
      summary: Логин пользователя
      operationId: signIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Неверные данные
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: JWT токены
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtAuthenticationDto"
...
  /api/cards/admin:
    get:
      tags:
      - Cards
      summary: Получить все карты (только ADMIN)
      description: Просмотр всех карт в системе с фильтрацией
      operationId: getAllCards
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - NOT_ACTIVE
          - ACTIVE
          - BLOCKED
          - EXPIRED
      - name: expirationBefore
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав ADMIN
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Список карт
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardDto"
      security:
      - bearerAuth: []
...
  /api/card-requests:
    get:
      tags:
      - Card Requests
      summary: Получить свои заявки
      description: Список всех заявок пользователя с пагинацией
      operationId: getMyRequests
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав USER
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Список заявок
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardRequestDto"
      security:
      - bearerAuth: []
  /api/card-requests/admin:
    get:
      tags:
      - Card Requests
      summary: Получить все заявки (ADMIN)
      description: Получение всех заявок с динамической фильтрацией и пагинацией
      operationId: getAllRequests
      parameters:
      - name: userId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: cardId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: type
        in: query
        required: false
        schema:
          type: string
          enum:
          - ACTIVATE
          - BLOCK
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - APPROVED
          - REJECTED
          - PENDING
      - name: requestedBalance
        in: query
        required: false
        schema:
          type: number
      - name: createdFrom
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: createdTo
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Нет прав ADMIN
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Список заявок
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardRequestDto"
      security:
      - bearerAuth: []
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    CardDto:
      type: object
      description: Информация о банковской карте
      properties:
        id:
          type: integer
          format: int64
          description: ID карты
          example: 10
        publicNumber:
          type: string
          format: uuid
          description: Публичный номер карты
        ownerId:
          type: integer
          format: int64
          description: ID владельца
          example: 1
        cardNumber:
          type: string
          description: Номер карты
          example: '**** **** **** 1234'
        expirationDate:
          type: string
          format: date
          description: Дата окончания действия карты
          example: 2028-12-31
        status:
          type: string
          description: Статус карты
          enum:
          - NOT_ACTIVE
          - ACTIVE
          - BLOCKED
          - EXPIRED
          example: ACTIVE
        balance:
          type: number
          description: Текущий баланс карты
          example: 750.5
    CardRequestBlockDto:
      type: object
      description: Запрос на блокировку карты
      properties:
        cardId:
          type: integer
          format: int64
          description: ID карты
          example: 10
    CardRequestDto:
      type: object
      description: Заявка пользователя на операцию с картой
      properties:
        id:
          type: integer
          format: int64
          description: ID заявки
          example: 100
        userId:
          type: integer
          format: int64
          description: ID пользователя
          example: 42
        cardId:
          type: integer
          format: int64
          description: ID карты
          example: 10
        type:
          type: string
          description: Тип запроса
          enum:
          - ACTIVATE
          - BLOCK
          example: BLOCK
        requestedBalance:
          type: number
          description: Запрашиваемый баланс
          example: 0
        status:
          type: string
          description: Статус заявки
          enum:
          - APPROVED
          - REJECTED
          - PENDING
          example: PENDING
        createdAt:
          type: string
          format: date-time
          description: Дата создания заявки
          example: 2026-01-15T14:30:00
    CardRequestActivateDto:
      type: object
      description: Запрос на активацию карты
      properties:
        requestedBalance:
          type: number
          description: Запрашиваемый баланс (опционально)
          example: 0
    RefreshTokenDto:
      type: object
      description: DTO для обновления access токена
      properties:
        refreshToken:
          type: string
          description: Refresh JWT токен
          example: eyJhbGciOiJIUzI1NiJ9.eyJ0eXBlIjoicmVmcmVzaCIsImV4cCI6MTcwMDAwMDAwMH0.signature
    JwtAuthenticationDto:
      type: object
      description: JWT токены после успешной аутентификации
      properties:
        token:
          type: string
          description: Access JWT токен
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMjMiLCJleHAiOjE2OTk5OTk5OTl9.signature
        refreshToken:
          type: string
          description: Refresh JWT токен
          example: eyJhbGciOiJIUzI1NiJ9.eyJ0eXBlIjoicmVmcmVzaCIsImV4cCI6MTcwMDAwMDAwMH0.signature
    LoginDto:
      type: object
      description: Данные для аутентификации пользователя
      properties:
        email:
          type: string
          format: email
          description: Email пользователя
          example: user@example.com
          minLength: 1
        password:
          type: string
          description: Пароль пользователя
          example: strongPassword123
          maxLength: 50
          minLength: 6
      required:
      - email
      - password
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        paged:
          type: boolean
        sort:
          $ref: "#/components/schemas/SortObject"
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    PageCardDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardDto"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageCardRequestDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/CardRequestDto"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT